+++ modules/features/home-manager/terminal/starship.nix
@@ -1,0 +1,12 @@
+{ self, inputs, ... }:
+
+{
+    flake.homeModules = {
+        starship = { config, ... }:
+            {
+                programs.starship = {
+                    enable = true;
+                };
+            };
+    };
+}