+++ modules/hosts/caelestia/configuration.nix
@@ -2,21 +2,19 @@
 
 {
     flake.nixosModules = {
-        illogical-impulseConfiguration = { config, ... }:
+        caelestiaConfiguration = { config, ... }:
             {
                 imports = [
                     self.nixosModules.core
-                    self.nixosModules.illogical-impulseHardware
-                    self.nixosModules.illogical-impulseHome
+                    self.nixosModules.caelestiaHardware
+                    self.nixosModules.caelestiaHome
                     self.nixosModules.nvidia
                     self.nixosModules.sddm
                     self.nixosModules.ssh
                     self.nixosModules.steam
                     self.nixosModules.flatpak
                     self.nixosModules.hyprland
-                    self.nixosModules.geoclue2
                     self.nixosModules.fileCompression-Decompression
-                    self.nixosModules.dconf
                 ];
             };
     };